Change my homework diretory
