+++ group04/564451732/hw1/src/Test.java
@@ -1,0 +1,8 @@
+
+public class Test {
+
+	public static void main(String[] args) {
+		
+	}
+
+}
