+++ group04/564451732/hw1/src/hw1/Iterator.java
@@ -1,0 +1,7 @@
+package hw1;
+
+public interface Iterator {
+	public boolean hasNext();
+	public Object next();
+
+}
