+++ group04/564451732/hw1/src/hw1/QueueImpl.java
@@ -1,0 +1,21 @@
+package hw1;
+
+public class QueueImpl {
+	LinkedList queueList = new LinkedList();
+	
+	public void enQueue(Object o){
+		queueList.add(o);
+	}
+	
+	public Object deQueue(){
+		return queueList.removeFirst();
+	}
+	
+	public boolean isEmpty(){
+		return queueList.size() == 0;
+	}
+	
+	public int size(){
+		return queueList.size();
+	}
+}
