--- group04/564451732/src/Test.java
@@ -1,8 +1,0 @@
-
-public class Test {
-
-	public static void main(String[] args) {
-		
-	}
-
-}
