--- group04/564451732/src/hw1/Iterator.java
@@ -1,7 +1,0 @@
-package hw1;
-
-public interface Iterator {
-	public boolean hasNext();
-	public Object next();
-
-}
