--- group04/564451732/src/hw1/List.java
@@ -1,9 +1,0 @@
-package hw1;
-
-public interface List {
-	public void add(Object o);
-	public void add(int index, Object o);
-	public Object get(int index);
-	public Object remove(int index);
-	public int size();
-}
